# back/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Autentication Users",
        "descriptions": "Serviço com foco em validações e gerenciamente de permissões de usuarios",
        "contact": {
            "email": "nivaldeir-silva@hotmail.com"
        }
    },
    "paths": {
...
        "/role": {
            "post": {
                "summary": "Cadastro de cargo",
                "descriptions": "",
                "tags": ["Role"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Role"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "name": "admin"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Já existe informação com esse...",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Busca de Roles",
                "descriptions": "",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/RoleArray"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}": {
            "get": {
                "descriptions": "Busca de role pelo ID",
                "summary": "Busca de role pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "descriptions": "Atualizando um role pelo ID",
                "summary": "Atualizando um role pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do role",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "name": "Admin"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "descriptions": "Deletando um usuario pelo ID",
                "summary": "Deletando um usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da role",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}/adding-permission": {
            "put": {
                "descriptions": "Adicionando uma permissão do cargo pelo IDD",
                "summary": "Adicionando uma permissão do cargo pelo IDD",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do role",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "permission": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}/deleting-permission": {
            "put": {
                "descriptions": "Deletando uma permissão do cargo pelo ID",
                "summary": "Deletando uma permissão do cargo pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Role"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do role",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "permission": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string"
                            },
                            "enabled": {
                                "type": "boolean"
                            },
                            "id": {
                                "type": "string"
                            },
                            "roleId": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "salt": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
            "RoleArray": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "boolean"
                                },
                                "permission": {
                                    "type": "array",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "required": ["message", "data"]
            },
            "Role": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "boolean"
                            },
                            "permission": {
                                "type": "array",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}